+++ src/core/libmath.h
@@ -45,7 +45,7 @@ QMap<qreal, PeakPick::MultiRegression> LeastSquares(const QVector<qreal>& x, con
 
 qreal SimpsonIntegrate(qreal lower, qreal upper, std::function<qreal(qreal, const QVector<qreal>)> function, const QVector<qreal>& parameter);
 
-qreal Stddev(const QVector<qreal>& vector, int end = 0);
+qreal Stddev(const QVector<qreal>& vector, int end = 0, double average = 0);
 
 namespace Cubic {
 qreal f(qreal x, qreal a, qreal b, qreal c, qreal d);