+++ Changelog.md
@@ -1,5 +1,6 @@
 # SupraFit 2.0 pre-Alpha
 
+- add some more comparison to reduction analysis
 - exported charts have now a transparent background and are croped to the opaque are (remove transparent bounding box)
 - some more colors and titles in statistics output
 - add optional in chart annotation as legend